requirements.txt
fastapi>=0.115
uvicorn>=0.30
numpy>=1.26
pandas>=2.2
faiss-cpu>=1.8
sentence-transformers>=2.7
pypdf>=4.2
python-docx>=1.1
openpyxl>=3.1